refactor: pr 요구사항 수정
