+++ src/main/java/com/ureca/uplait/domain/mypage/controller/MyPageController.java
@@ -1,14 +1,14 @@
 package com.ureca.uplait.domain.mypage.controller;
 
-import com.ureca.uplait.domain.mypage.dto.MyPageResponse;
-import com.ureca.uplait.domain.mypage.dto.MyPageUpdateRequest;
-import com.ureca.uplait.domain.mypage.dto.MyPageUpdateResponse;
-import com.ureca.uplait.domain.mypage.dto.MyReviewsResponse;
+import com.ureca.uplait.domain.mypage.dto.response.MyPageResponse;
+import com.ureca.uplait.domain.mypage.dto.request.MyPageUpdateRequest;
+import com.ureca.uplait.domain.mypage.dto.response.MyPageUpdateResponse;
+import com.ureca.uplait.domain.mypage.dto.response.MyReviewsResponse;
 import com.ureca.uplait.domain.mypage.service.MyPageService;
 import com.ureca.uplait.domain.user.entity.User;
+import com.ureca.uplait.global.response.CommonResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
@@ -21,26 +21,23 @@ public class MyPageController {
 
     @Operation(summary = "마이페이지 조회", description = "개인정보 조회: 로그인 필요")
     @GetMapping("user/detail")
-    public ResponseEntity<MyPageResponse> getMyPage(@AuthenticationPrincipal User user) {
+    public CommonResponse<MyPageResponse> getMyPage(@AuthenticationPrincipal User user) {
 
-        MyPageResponse myPageResponse = myPageService.getMyPage(user.getId());
-        return ResponseEntity.ok(myPageResponse);
+        return CommonResponse.success(myPageService.getMyPage(user.getId()));
     }
 
     @Operation(summary = "마이페이지 개인정보 수정 ", description = "개인정보 수정 후 수정 완료 버튼을 누름: 로그인 필요")
     @PutMapping("user/detail/update")
-    public ResponseEntity<MyPageUpdateResponse> updateMyPage(@AuthenticationPrincipal User user
+    public CommonResponse<MyPageUpdateResponse> updateMyPage(@AuthenticationPrincipal User user
             , @RequestBody MyPageUpdateRequest myPageUpdateRequest) {
 
-        MyPageUpdateResponse myPageUpdateResponse = myPageService.updateMyPage(user, myPageUpdateRequest);
-        return ResponseEntity.ok(myPageUpdateResponse);
+        return CommonResponse.success(myPageService.updateMyPage(user, myPageUpdateRequest));
     }
 
     @Operation(summary = "내가 쓴 리뷰 목록", description = "내가 쓴 리뷰 목록 확인 : 로그인 필요")
     @GetMapping("user/detail/review")
-    public ResponseEntity<List<MyReviewsResponse>> getMyReview(@AuthenticationPrincipal User user) {
+    public CommonResponse<List<MyReviewsResponse>> getMyReview(@AuthenticationPrincipal User user) {
 
-        List<MyReviewsResponse> reviewsResponse = myPageService.getMyReview(user.getId());
-        return ResponseEntity.ok(reviewsResponse);
+        return CommonResponse.success(myPageService.getMyReview(user.getId()));
     }
 }
