+++ src/main/java/com/ureca/uplait/domain/mypage/dto/request/MyPageUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.ureca.uplait.domain.mypage.dto;
+package com.ureca.uplait.domain.mypage.dto.request;
 
 import com.ureca.uplait.domain.user.entity.User;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -11,15 +11,10 @@ import lombok.Setter;
 @AllArgsConstructor
 @Schema(description = "마이페이지 수정 요청")
 public class MyPageUpdateRequest {
+    @Schema(description = "핸드폰 번호", example = "010-1234-5678")
     private String phoneNumber;
+    @Schema(description = "이메일", example = "ureca@kakao.com")
     private String email;
+    @Schema(description = "광고 수신 동의", example = "true")
     private boolean adAgree;
-
-    public static MyPageUpdateRequest from(User user) {
-        return new MyPageUpdateRequest(
-                user.getPhoneNumber()
-                , user.getEmail()
-                , user.getAdAgree()
-        );
-    }
 }
